requirements.txt
qrcode[pil]==6.1
pillow==7.1.2
numpy==1.18.5
imageio==2.8.0
scikit-image==0.16.2
matplotlib==3.2.1
pyzbar==0.1.8
tqdm==4.46.0
click==7.1.2
imgaug==0.4.0
pandas==1.0.4
scikit-learn==0.23.1
scipy==1.4.1